flesh out 'sunbeam config' command, cleanup install.sh
